Leet_code_palindrome.py
In this exercise i attempted to solve the leet code palindrome problem, it was easier than the roman numeral one however its was still slightly challenging. The listNode object its used to feed in data that is added to the palindrome list. after the palindrome list is reversed it is numerically compared to itself and if it is equal then it returns true